pyproject.toml
[project]
name = "pdf-extractor"
version = "0.1.0"
description = "Extract Q&A pairs from math-heavy PDF documents into LaTeX"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    # PDF Processing
    "pymupdf>=1.24.0",
    "Pillow>=10.0.0",

    # LLM Clients
    "openai>=1.0.0",
    "anthropic>=0.40.0",

    # Data Handling
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",

    # CLI
    "typer>=0.12.0",
    "rich>=13.0.0",
]

[project.scripts]
pdf-extractor = "src.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
